address Gutos feedback
