+++ app/core/SnapKeyring/SnapKeyring.ts
@@ -2,50 +2,7 @@ import { SnapKeyring } from '@metamask/eth-snap-keyring';
 import type { SnapController } from '@metamask/snaps-controllers';
 import { SnapKeyringBuilderMessenger } from './types';
 import Logger from '../../util/Logger';
-import { SNAP_MANAGE_ACCOUNTS_CONFIRMATION_TYPES } from '../RPCMethods/RPCMethodMiddleware';
-
-interface CreateAccountConfirmationResult {
-  success: boolean;
-  name?: string;
-}
-
-/**
- * Show the account name suggestion confirmation dialog for a given Snap.
- *
- * @param snapId - Snap ID to show the account name suggestion dialog for.
- * @param controllerMessenger - The controller messenger instance.
- * @param accountNameSuggestion - Suggested name for the new account.
- * @returns The user's confirmation result.
- */
-export async function showAccountNameSuggestionDialog(
-  snapId: string,
-  controllerMessenger: SnapKeyringBuilderMessenger,
-  accountNameSuggestion: string,
-): Promise<CreateAccountConfirmationResult> {
-  try {
-    const confirmationResult = (await controllerMessenger.call(
-      'ApprovalController:addRequest',
-      {
-        origin: snapId,
-        type: SNAP_MANAGE_ACCOUNTS_CONFIRMATION_TYPES.showNameSnapAccount,
-        requestData: {
-          snapSuggestedAccountName: accountNameSuggestion,
-        },
-      },
-      true,
-    )) as CreateAccountConfirmationResult;
-
-    if (confirmationResult) {
-      return {
-        success: confirmationResult.success,
-        name: confirmationResult.name,
-      };
-    }
-    return { success: false };
-  } catch (e) {
-    throw new Error(`Error occurred while showing name account dialog.\n${e}`);
-  }
-}
+import { showAccountNameSuggestionDialog } from './utils/showDialog';
 
 /**
  * Constructs a SnapKeyring builder with specified handlers for managing snap accounts.
@@ -93,7 +50,6 @@ export const snapKeyringBuilder = (
         snapId: string,
         handleUserInput: (accepted: boolean) => Promise<void>,
         accountNameSuggestion = '',
-        _displayConfirmation = false,
       ) => {
         const { id: addAccountFlowId } = controllerMessenger.call(
           'ApprovalController:startFlow',
